--- a/src/mscorlib/src/System/Threading/SemaphoreSlim.cs
+++ b/src/mscorlib/src/System/Threading/SemaphoreSlim.cs
@@ -58,16 +58,6 @@ namespace System.Threading
         // threading and decrements it back after that. It is used as flag for the release call to know if there are
         // waiting threads in the monitor or not.
         private volatile int m_waitCount;
-        
-        // The number of threads entered Wait method. 
-        // Increments every time before calling Monitor.Enter() inside Wait method.
-        // Allow to calculate the number of threads that are about to block on m_lockObj.
-        private int m_enteringWaiterNumber;
-
-        // The number of threads exited Wait method.
-        // Increments every time before calling Monitor.Exit() inside Wait method.
-        // Allow to calculate the number of threads that are about to block on m_lockObj.
-        private volatile int m_finishedWaiterNumber;
 
         // Dummy object used to in lock statements to protect the semaphore count, wait handle and cancelation
         private object m_lockObj;
@@ -338,7 +328,6 @@ namespace System.Threading
             bool waitSuccessful = false;
             Task<bool> asyncWaitTask = null;
             bool lockTaken = false;
-            bool enteringWaiterNumberUpdated = false;
 
             //Register for cancellation outside of the main lock.
             //NOTE: Register/deregister inside the lock can deadlock as different lock acquisition orders could
@@ -346,34 +335,15 @@ namespace System.Threading
             CancellationTokenRegistration cancellationTokenRegistration = cancellationToken.InternalRegisterWithoutEC(s_cancellationTokenCanceledEventHandler, this);
             try
             {
-                // Perf: trying to wait with spin.
+                // Perf: first spin wait for the count to be positive, but only up to the first planned yield.
                 //       This additional amount of spinwaiting in addition
                 //       to Monitor.Enter()’s spinwaiting has shown measurable perf gains in test scenarios.
                 //
-
-                // Check the conditions to perform spinwaiting
-                //  1. m_currentCount is less than the number of entered threads (can wait for Release)
-                //  2. number of waiters is less than core count (otherwise the spinning gives nothing)
-                if (m_currentCount < 1 + Volatile.Read(ref m_enteringWaiterNumber) - m_finishedWaiterNumber && m_waitCount <= PlatformHelper.ProcessorCount)
+                SpinWait spin = new SpinWait();
+                while (m_currentCount == 0 && !spin.NextSpinWillYield)
                 {
-                    // Calculating Id for current thread
-                    int myEntryId = 0;
-                    try { }
-                    finally
-                    {
-                        myEntryId = Interlocked.Increment(ref m_enteringWaiterNumber);
-                        enteringWaiterNumberUpdated = true;
-                    }
-
-                    SpinWait spin = new SpinWait();
-                    // This condition allows to arrange all threads in the order of their appearance
-                    // ('myEntryId - m_finishedWaiterNumber - 1' is the number of waiting threads entered before current)
-                    while (m_currentCount < myEntryId - m_finishedWaiterNumber && !spin.NextSpinWillYield)
-                    {
-                        spin.SpinOnce();
-                    }
+                    spin.SpinOnce();
                 }
-                
                 // entering the lock and incrementing waiters must not suffer a thread-abort, else we cannot
                 // clean up m_waitCount correctly, which may lead to deadlock due to non-woken waiters.
                 try { }
@@ -444,20 +414,6 @@ namespace System.Threading
             }
             finally
             {
-                if (enteringWaiterNumberUpdated)
-                {
-                    if (lockTaken)
-                    {
-                        m_finishedWaiterNumber++;
-                    }
-                    else
-                    {
-                        // Incrementing of m_finishedWaiterNumber can cause the 'lost update' situation, 
-                        // so it is better to decrement atomically m_enteringWaiterNumber
-                        Interlocked.Decrement(ref m_enteringWaiterNumber);
-                    }
-                }
-                
                 // Release the lock
                 if (lockTaken)
                 {