@brianrob 2015-02-09
@ikopylov, thank you very much for your contribution.

I have not had a chance to look at your change yet.  I'll take a look and reply back with comments.

Thanks.