@ikopylov 2015-02-09
Unfortunately this tweak doesn't solve the problem for all configurations.

I have an opportunity to test it on 24 core CPU (2x Intel Xeon X5675) and here's the results:
Original:
```
1, 1:     280ms
24, 24:   558ms
1, 8:     526ms
1, 12:   4201ms
1, 24:  76981ms
4, 24:   1093ms
4, 48:  33501ms
8, 1:     184ms
12, 1:    241ms
24, 1:    248ms
```

With tweak:
```
1, 1:     202ms
24, 24:   572ms
1, 8:     455ms
1, 12:  21572ms
1, 24:  92623ms
4, 24:  11188ms
4, 48:  10140ms
8, 1:     183ms
12, 1:    251ms
24, 1:    251ms
```
*Because of the slow execution I've reduced the number of iterations by factor of 10.


As you can see the updated version of SemaphoreSlim works inferior to the original in this environment. Now it's obvious that this issue happened not because the number of threads more than number of cores. But still there's definitely some problem with the implementation of SemaphoreSlim. 

I tried to wrap Wait call in additional lock and it begins to work as it should (OMG).
So with the code of test that looks like this:
```C#
while (!srcCancel.IsCancellationRequested)
{
    lock (waiterSyncObj)
        sem.Wait(myToken);
    Thread.SpinWait(takeSpin);
}
```
I have following results:
```
1, 1:   253ms
24, 24: 374ms
1, 8:   244ms
1, 12:  375ms
1, 24:  405ms
4, 24:  330ms
4, 48:  324ms
8, 1:   176ms
12, 1:  236ms
24, 1:  247ms
```

From now I think that the problem caused by the call of Monitor.PulseAll. It lead to awakening of all threads, but only few of them exits the lock, all others come back to Wait state.
I'll continue investigation, but this Pull-Request should not be accepted at this state.