@brianrob 2015-02-09
@ikopylov, thank you very much for your contribution.

In general, it sounds like you're doing a great job of thinking about and testing a wide variety of configurations.  This is great.

When it comes to spinning and locking changes, we've generally spent a great deal of time building up confidence in the changes.  So, while I don't want to discourage you from making changes to help improve things in this space, I do want you to be aware that the bar for changes here will be very high, and that you'll likely get lots of questions and or requests to test various configs.