@ericeil 2015-02-10
@ikopylov, thanks for looking into this.  What you have implemented is effectively a "ticket lock" (so named because of its resemblance to "now serving" tickets you might find at a bakery, or the DMV).  This has the nice property of reducing memory contention while threads are waiting, as you intend.  However, it also has the possibly unintended consequence of making this a "fair" lock, in that requests are serviced in the order in which they were received.  This can lead to a number of performance issues, including so-called "lock convoys" as well as "thrashing" of the CPU caches:  the algorithm prefers to give control to threads that have been waiting the longest, and therefore are least likely to have their data "hot" in the CPU cache.  These effects tend to be magnified as more threads contend for the lock, so an algorithm that works very well for low concurrency levels (one or two threads per core) may scale very badly as more threads are added.  I don't know for sure that either of these are the cause of the performance issues you have observed, but perhaps this provides some clues.

I agree that Montor.Pulse is a likely culprit in the existing implementation, and it may be worthwhile to look for alternatives.  

I'll also echo Brian's statement that the bar for changes in this area is very high.  Synchronization primitives need to perform reasonably well across a vast number of machine configurations and usage scenarios, and it is simply not practical to test *all* of the scenarios and configurations that may be important to users, so we are very careful about accepting changes to synchronization code that has been working "well enough" so far.