@ikopylov 2015-02-10
Thanks for your response. 

I understand that this is a critical part of Framework and all changes should be tested carefully in many different environments. But the performance drop here is enormous (up to 250x slower than expected), so I believe that it should be fixed as soon as possible. I pointed out the problem and if I fail to fix it, I hope that CoreCLR team will do it.

Meanwhile, I've tested the hypothesis that the performance drop connected with Monitor.PulseAll call. I've added a counter to estimate the number of false-wakeups and here the results:
```
1, 1:   1489ms.  False-WakeUps: 0.
1, 8:   3467ms.  False-WakeUps: 3599.
1, 16: 69707ms.  False-WakeUps: 4708503.
```

Now I'm pretty sure, that this is the root of the problem. My initial "ticket-lock" approach just hide it in some configurations.

Probably the easiest way to fix the issue is to call Monitor.Pulse only by the number of waiters.
So instead of this:
```C#
if (currentCount == 1 || waitCount == 1)
{
    Monitor.Pulse(m_lockObj);
}
else if (waitCount > 1)
{
    Monitor.PulseAll(m_lockObj);
}
```
Do something like this:
```C#
for (int i = 0; i < Math.Min(releaseCount, waitCount); i++)
{
    Monitor.Pulse(m_lockObj);
}
```

Quick test shows that this change improves the performance and reduce the number of false-wakeups, but the number of Monitor.Pulse call should be evaluated a little bit smarter.