@ikopylov 2015-02-11
I have update Pull Request according to my previous comment. In this modification Monitor.Pulse called by the number of relaseCount (and waiters) instead of calling Monitor.PulseAll. This approach reduce the number of False-Wakeups in corner cases.

Here the results of the test (8 core cpu Intel i7-4770k).
Original:
```
1, 1:    1491ms.   False-Wakeups:       0.
8, 8:    2293ms.   False-Wakeups:       0.
16, 16:  2390ms.   False-Wakeups:       5.
1, 8:    3443ms.   False-Wakeups:    3853. 
1, 16:  53624ms.   False-Wakeups: 3427029.
4, 24:   5915ms.   False-Wakeups:   74252.
8, 1:    1498ms.   False-Wakeups:       0.
16, 1:   1476ms.   False-Wakeups:       0.
```
Modified:
```
1, 1:    1440ms.   False-Wakeups:       0.
8, 8:    2346ms.   False-Wakeups:       0.
16, 16:  2345ms.   False-Wakeups:       2.
1, 8:    3475ms.   False-Wakeups:     503. 
1, 16:  18738ms.   False-Wakeups:   73251.
4, 24:   4607ms.   False-Wakeups:    7924.
8, 1:    1551ms.   False-Wakeups:       0.
16, 1:   1509ms.   False-Wakeups:       0.
```

Still not the expected results ("1, 16" should be no more than 2 times slower than "1, 8"), but much better than original.