@ikopylov 2015-02-12
Test on 24 core CPU (2x Intel Xeon X5675).
*The number of iterations reduced by factor of 10.
Original:
```
1, 1:      203ms.   False-wakeups:        0.
24, 24:    497ms.   False-wakeups:        2.
1, 12:    5077ms.   False-wakeups:   470355.
1, 24:   77203ms.   False-wakeups:  8311677.
1, 48:  161288ms.   False-wakeups: 19293180.
4, 48:   29413ms.   False-wakeups:  1811314.
24, 1:     251ms.   False-wakeups:        0. 
```
Modified:
```
1, 1:      198ms.   False-wakeups:        0.
24, 24:    519ms.   False-wakeups:        2.
1, 12:     787ms.   False-wakeups:     2900.
1, 24:    2705ms.   False-wakeups:    13404.
1, 48:    3863ms.   False-wakeups:    23313.
4, 48:     862ms.   False-wakeups:      384.
24, 1:     254ms.   False-wakeups:        0. 
```
Looks great.