@vancem 2015-02-12
The change looks simple which is great.    My concern is that we need an analysis that tells us that we won't have deadlock bugs because we did not release enough threads.   

Can you also describe the scenario that motivated the change (that is what is your scenario that has few Releasers but many waiters?